requirements.txt
streamlit==1.28.0
tensorflow==2.15.0
pillow==10.0.0
boto3==1.34.0
